+++ src/rust/src/single_cell/mod.rs
@@ -1,4 +1,6 @@
 pub mod harmony_gpu;
 pub mod knn_gpu;
 pub mod pca_gpu;
+pub mod sc_utils;
 pub mod scenic_gpu;
+pub mod seacells_gpu;